change directory and exec module
