+++ lib/client/openai.rb
