+++ lib/main.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+Resquirrel.generate_release_note
